Fixing the way we handle and ignore user exceptions. I thought I knew how this worked the first time I checked in this feature. Now, I actually do.
On Engine::SetRegistryRoot, We read default exception settings from the registry.
Calls to Engine::SetException, RemoveSetException, and RemoveAllSetExceptions
modify the overrides.

In the event callback for exceptions, we look in the overrides and then the defaults.

Also, if an exception isn't found in one of the predefined categories (D/Win32), then
we use the info for the catch-all root exception for the category it belongs to.

The exception settings in the registry files used obsolete state settings.
These are fixed.