fix gaps in the dmd/optlink line number debug information
optlink creates bad end offsets in the OMFSourceFile data: the end contains the offset of the last line number in the section, but not the last code byte. This causes the debugger to switch to assembly because there is a gap without associated source.

workaround: create a bitmap of all available line offsets plus some more offsets that are probable function boundaries, and expand the end offset up to the next found line offset.