add display of global or static variables without having to give the complete scope
this is done by extracting the current scope from the function function name, and then prepending it to the symbol. If not found, cut off the part after the last '.' and try again.

This does not work for functions with the extern(C) name-decoration, because these do not contain the scope.