Changing the way versions are managed.
Each executable now shows a product version of major.minor.
The commas that used to appear are now dots.

This is the full scheme for file versions: major.minor.build.revision
- major: big changes
- minor: little changes
- build: unused for now, set to 0
- revision: changes to the little changes (major.minor)

For a group of related check-ins, update major, minor, and revision as needed, and check it in.
When finished, make a tag of it: copy /trunk to /tags/<version>.