From the RemoteDebuggerProxy in the debug engine, make two RPC contexts for each session: - one for commands from the current event callback thread - one for commands from all other threads
That way, commands meant to run in, and stemming from an event fired from the event callback thread in MagoRemote can always be dispatched, and are never blocked by a command from another thread. Deadlock is avoided.

In order to bring this about, allow binding more than one context to the same session data in MagoRemote.